revise UnmaskedNanCounter as maskNans (#2951)
NANs that are already masked weren't getting identified as needing to be
interpolated over (because of the check for _pbMask in the apply loop),
and therefore aren't removed.  In fact, we want to remove all NANs, not
just the ones that are unmasked (just because they have a mask bit set
doesn't mean they will be interpolated over by some other function, and
we want to remove all NANs).

* Made UnmaskedNanCounter a function (because there's no need to preserve
  state) and updated the name to more closely match the functionality.
* Made the mask bits to allow and mask configurable.
* Updated docs.
* Removed instantiation for int and uint16_t (have no NAN).  We could
  introduce noop specialisations for these if desired.
* Added test.