Consolidate ISR interpolation of bad pixels.
Previously, masking and interpolation was done on a by-defect-type
basis: a set of bad pixels were masked and interpolated, then the
process repeated on different bad pixels.  However, this ignores the
case where different types of bad pixels may be contiguous, resulting
in interpolated pixels being derived based on previously interpolated
values.

This ticket reorders ISR processing such that all bad pixels are now
masked prior to the brighter-fatter correction.  A temporary
interpolation is done to run brighter-fatter, but only to derive the
expected correction for the pixels.  The remaining ISR stages are run
on the pre-interpolated pixels, with the final interpolation performed
only after all other pixel modifications are finished.  The
pre-interpolated exposure is retained and written out if the
doSaveInterpPixels configuration option is set to True.